tests/requirements.txt
-r ../requirements.txt

coverage>=5.5
pytest-cov>=2.12.1
pytest-flask>=1.2.0
responses>=0.13.3
psycopg2-binary>=2.9.3